ファイル番号8のnorm対応
