+++ ForSubmit/srcs/8_1_free_and_destroy.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   8_free_and_destroy.c                               :+:      :+:    :+:   */
+/*   8_1_free_and_destroy.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yokitaga <yokitaga@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/20 23:43:54 by yokitaga          #+#    #+#             */
-/*   Updated: 2023/11/25 15:18:36 by yokitaga         ###   ########.fr       */
+/*   Updated: 2023/11/25 16:50:14 by yokitaga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,16 +76,4 @@ void	free_pos_dir_plane(t_data *data)
 		free(data->dir);
 	if (data->plane != NULL)
 		free(data->plane);
-}
-
-void	free_and_destroy_all(t_data *data)
-{
-	destroy_textures(data);
-	free_map_data_and_path(data);
-	mlx_destroy_window(data->mlx, data->mlx_win);
-	mlx_destroy_display(data->mlx);
-	free_texture(data);
-	free(data->mlx);
-	free_pos_dir_plane(data);
-	free(data);
 }
